added srctype step and fixed the name and ordering of the tests according to the calwebb_spec2.py script
