--- calwebb_spec2_pytests/B_assign_wcs/test_assign_wcs.py
@@ -1,63 +1,0 @@
-
-"""
-py.test module for unit testing the assign_wcs step.
-"""
-
-import pytest
-import os
-from jwst.assign_wcs.assign_wcs_step import AssignWcsStep
-
-from .. import core_utils
-from . import assign_wcs_utils
-
-
-# Set up the fixtures needed for all of the tests, i.e. open up all of the FITS files
-
-# Default names of pipeline input and output files
-@pytest.fixture(scope="module")
-def set_inandout_filenames(request, config):
-    step = "assign_wcs"
-    step_info = core_utils.set_inandout_filenames(step, config)
-    step_input_filename, step_output_filename, in_file_suffix, out_file_suffix, True_steps_suffix_map = step_info
-    return step, step_input_filename, step_output_filename, in_file_suffix, out_file_suffix, True_steps_suffix_map
-
-
-# fixture to read the output file header
-@pytest.fixture(scope="module")
-def output_hdul(set_inandout_filenames, config):
-    set_inandout_filenames_info = core_utils.read_info4outputhdul(config, set_inandout_filenames)
-    step, txt_name, step_input_file, step_output_file, run_calwebb_spec2, outstep_file_suffix = set_inandout_filenames_info
-    stp = AssignWcsStep()
-    # if run_calwebb_spec2 is True calwebb_spec2 will be called, else individual steps will be ran
-    step_completed = False
-    if not run_calwebb_spec2:
-        if config.getboolean("steps", step):
-            print ("*** Step "+step+" set to True")
-            if os.path.isfile(step_input_file):
-                #result = stp.call(step_input_file)
-                #result.save(step_output_file)
-                step_completed = True
-                core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
-                hdul = core_utils.read_hdrfits(step_output_file, info=False, show_hdr=False)
-                return hdul
-            else:
-                print("Skipping step. Intput file "+step_input_file+" does not exit.")
-                core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
-                pytest.skip("Skipping "+step+" because the input file does not exist.")
-        else:
-            core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
-            pytest.skip("Skipping "+step+". Step set to False in configuration file.")
-
-
-
-# Unit tests
-
-def test_wavstart_exists(output_hdul):
-    assert assign_wcs_utils.wavstart_exists(output_hdul), "The keyword WAVSTART was not added to the header."
-
-def test_sporder_exists(output_hdul):
-    assert assign_wcs_utils.sporder_exists(output_hdul), "The keyword SPORDER was not added to the header."
-
-def test_s_wcs_exists(output_hdul):
-    assert assign_wcs_utils.s_wcs_exists(output_hdul), "The keyword S_WCS was not added to the header --> extract_2d step was not completed."
-
