+++ calwebb_spec2_pytests/D_msa_flagging/test_msa_flagging.py
@@ -27,25 +27,32 @@ def set_inandout_filenames(request, config):
 def output_hdul(set_inandout_filenames, config):
     set_inandout_filenames_info = core_utils.read_info4outputhdul(config, set_inandout_filenames)
     step, txt_name, step_input_file, step_output_file, run_calwebb_spec2, outstep_file_suffix = set_inandout_filenames_info
-    stp = MSAFlagOpenStep()
-    # if run_calwebb_spec2 is True calwebb_spec2 will be called, else individual steps will be ran
-    step_completed = False
-    if not run_calwebb_spec2:
-        if config.getboolean("steps", step):
-            print ("*** Step "+step+" set to True")
-            if os.path.isfile(step_input_file):
-                result = stp.call(step_input_file)
-                result.save(step_output_file)
-                step_completed = True
-                core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
-                hdul = core_utils.read_hdrfits(step_output_file, info=False, show_hdr=False)
-                return hdul
+    skip_runing_pipe_step = config.getboolean("tests_only", "_".join((step, "tests")))
+    # Only run step if data is MOS
+    inhdu = core_utils.read_hdrfits(step_input_file, info=False, show_hdr=False)
+    if core_utils.check_MOS_true(inhdu):
+        stp = MSAFlagOpenStep()
+        # if run_calwebb_spec2 is True calwebb_spec2 will be called, else individual steps will be ran
+        step_completed = False
+        if not run_calwebb_spec2:
+            if config.getboolean("steps", step):
+                print ("*** Step "+step+" set to True")
+                if os.path.isfile(step_input_file):
+                    if not skip_runing_pipe_step:
+                        result = stp.call(step_input_file)
+                        result.save(step_output_file)
+                    step_completed = True
+                    core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
+                    hdul = core_utils.read_hdrfits(step_output_file, info=False, show_hdr=False)
+                    return hdul
+                else:
+                    core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
+                    pytest.skip("Skipping "+step+" because the input file does not exist.")
             else:
                 core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
-                pytest.skip("Skipping "+step+" because the input file does not exist.")
-        else:
-            core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
-            pytest.skip("Skipping "+step+". Step set to False in configuration file.")
+                pytest.skip("Skipping "+step+". Step set to False in configuration file.")
+    else:
+        pytest.skip("Skipping "+step+" because data is not MOS.")
