+++ calwebb_spec2_pytests/K_resample/test_resample.py
@@ -27,6 +27,7 @@ def set_inandout_filenames(request, config):
 def output_hdul(set_inandout_filenames, config):
     set_inandout_filenames_info = core_utils.read_info4outputhdul(config, set_inandout_filenames)
     step, txt_name, step_input_file, step_output_file, run_calwebb_spec2, outstep_file_suffix = set_inandout_filenames_info
+    skip_runing_pipe_step = config.getboolean("tests_only", "_".join((step, "tests")))
     # Only run step if data is not IFU
     inhdu = core_utils.read_hdrfits(step_input_file, info=False, show_hdr=False)
     if not core_utils.check_IFU_true(inhdu):
@@ -37,8 +38,9 @@ def output_hdul(set_inandout_filenames, config):
             if config.getboolean("steps", step):
                 print ("*** Step "+step+" set to True")
                 if os.path.isfile(step_input_file):
-                    result = stp.call(step_input_file)
-                    result.save(step_output_file)
+                    if not skip_runing_pipe_step:
+                        result = stp.call(step_input_file)
+                        result.save(step_output_file)
                     step_completed = True
                     core_utils.add_completed_steps(txt_name, step, outstep_file_suffix, step_completed)
                     hdul = core_utils.read_hdrfits(step_output_file, info=False, show_hdr=False)
