+++ calwebb_spec2_pytests/cwspec2_config.cfg
@@ -8,7 +8,6 @@
 # input_file = name of file to be tested
 [calwebb_spec2_input_file]
 working_directory = /Users/pena/Documents/PyCharmProjects/nirspec/pipeline/src/sandbox/zzzz
-#data_directory = /Users/pena/Documents/PyCharmProjects/nirspec/pipeline/build7/test_data/MOS_CV3/complete_pipeline_testset
 data_directory = /Users/pena/Documents/PyCharmProjects/nirspec/pipeline/src/sandbox/zzzz
 # name of the suffix map (will be saved in the pytests directory)
 True_steps_suffix_map = True_steps_suffix_map.txt
@@ -35,13 +34,14 @@ fflat_path = /Users/pena/Documents/PyCharmProjects/nirspec/pipeline/src/sandbox/
 # this file has to contain an output file name for each step with the expected step suffixes (each
 # suffix is added to the previous file name, e.g. bkg_subtract step; if a step is skipped so is the
 # corresponding suffix, the suffix map will be saved in file True_steps_suffix_map.txt in the
-# pytests directory):
-#   bkg_subtract  -->  *_rate_subtract_images.fits
-#   assign_wcs  -->  *_subtract_images_assign_wcs.fits
+# pytests directory). The final name will be long but it is clear what steps where ran:
+#   assign_wcs  -->  *_assign_wcs.fits
+#   bkg_subtract  -->  *_subtract_images.fits
 #   imprint_subtract  -->  *_imprint.fits
 #   msa_flagging  -->  *_msa_flag.fits
 #   extract_2d  -->  *_extract_2d.fits
 #   flat_field  -->  *_flat_field.fits
+#   srctype  -->  No fits file created
 #   straylight  -->  *_straylight.fits (MIRI)
 #   fringe  -->  *_fringe.fits (MIRI)
 #   pathloss  -->  *_pathloss.fits
@@ -57,18 +57,37 @@ calwebb_spec2_cfg = /Users/pena/Documents/PyCharmProjects/nirspec/pipeline/build
 # steps to run (the order in which they will be ran is set by the pytests)
 # In order to skip a step set the value to False
 [steps]
-bkg_subtract = False
 assign_wcs = True
+bkg_subtract = False
 imprint_subtract = False
 msa_flagging = False
 extract_2d = True
 flat_field = True
+srctype = True
 pathloss = True
 barshadow = True
 photom = True
 resample_spec = False
-cube_build = False
+cube_build = True
 extract_1d = False
+
+# This section allows to run the pytests skipping a successfull re-run of the pipeline.
+# If a variable in this block is set to True, then the pipeline will not run, but the step is assumed to be
+# completed, and the output file is expected to be in the working directory with the name in the name in the map.
+[tests_only]
+assign_wcs_tests = True
+bkg_subtract_tests = False
+imprint_subtract_tests = False
+msa_flagging_tests = False
+extract_2d_tests = True
+flat_field_tests = True
+srctype_tests = False
+pathloss_tests = False
+barshadow_tests = False
+photom_tests = False
+resample_spec_tests = False
+cube_build_tests = False
+extract_1d_tests = False
 
 # additional arguments needed for specific steps
 [additional_arguments]
